Remove the 'A' type parameter from 'OAuth2<A, C>'.
'A' was only a type parameter so that static dispatch could be used,
which is not worth the ergonomic penalty of specifying it everywhere nor
the confusing ability to register the same callback twice with different
adapters.

The 'C' type parameter is still useful, allowing multiple callbacks to
be used in one application without using an out-of-band key to
differentiate them.